Fix visualization bugs and add overlay

+++ fast_track/object_detection/object_detector.py
@@ -32,7 +32,7 @@ class ObjectDetector(metaclass=ABCMeta):
 
         # Generate class colors for detection visualization
         rng = np.random.default_rng()
-        self.class_colors = [rng.integers(low=0, high=255, size=3).tolist() for _ in self.names]
+        self.class_colors = [rng.integers(low=0, high=255, size=3, dtype=np.uint8).tolist() for _ in self.names]
 
     @property
     def image_width(self) -> int:
